Add files via upload
o apply data preprocessing and preparation techniques in order to obtain clean data. To build machine learning models able to predict house prices based on different features. To analyze and compare the model's performance in order to choose the best model.

Skills and Tools

Cluster Analysis, Decision Trees, Logistic Regression, Neural Networks, Python

Conclusion

In this project, we train our models on the basis of two years of data on sale patterns according to the different conditions. To get the perfect model we choose different machine learning techniques like Random Forest, MLR, KNN, etc. We get an RMSE = 0.11 and an Accuracy = 80% for the Gradient Boosting Method which happens to be our best-fitted model in this case. It helps people to buy a house in budget an reduce loss of money.